+++ pages/src/controller/controller.php
@@ -16,5 +16,10 @@ class Controller{
 		}
 		return $validate;
 	}
+	public function backup(){
+		require_once "model/model.php";
+		$model = new Model(null);
+		return $model->backup();
+	}
 }
 ?>