+++ pages/src/link-file/foot-js.php
@@ -63,7 +63,7 @@
           </div>
         </div>
         <div class="modal-footer" style="background-color:lightblue !important">
-          <button type="reset" class="btn btn-primary pull-left" data-dismiss="modal">Close</button>
+          <button type="reset" class="btn btn-primary pull-left" id="fac_close_btn">Close</button>
           <button type="submit" class="btn btn-primary">Save</button>
         </div>
       </div>
@@ -75,33 +75,26 @@
 
 
 <!-- jQuery 2.2.3 -->
-<script src="../../../plugins/jQuery/jquery-2.2.3.min.js"></script>
+<script src="../../plugins/jQuery/jquery-2.2.3.min.js"></script>
 <!-- Bootstrap 3.3.6 -->
-<script src="../../../bootstrap/js/bootstrap.min.js"></script>
+<script src="../../bootstrap/js/bootstrap.min.js"></script>
 <!-- FastClick -->
-<script src="../../../plugins/fastclick/fastclick.js"></script>
+<script src="../../plugins/fastclick/fastclick.js"></script>
 <!-- AdminLTE App -->
-<script src="../../../dist/js/app.min.js"></script>
+<script src="../../dist/js/app.min.js"></script>
 <!-- AdminLTE for demo purposes -->
-<script src="../../../dist/js/demo.js"></script>
+<script src="../../dist/js/demo.js"></script>
 
 <!-- DataTables -->
-<script src="../../../plugins/datatables/jquery.dataTables.min.js"></script>
-<script src="../../../plugins/datatables/dataTables.bootstrap.min.js"></script>
+<script src="../../plugins/datatables/jquery.dataTables.min.js"></script>
+<script src="../../plugins/datatables/dataTables.bootstrap.min.js"></script>
 <!-- SlimScroll -->
-<script src="../../../plugins/slimScroll/jquery.slimscroll.min.js"></script>
-<script src="../../../plugins/iCheck/icheck.min.js"></script>
+<script src="../../plugins/slimScroll/jquery.slimscroll.min.js"></script>
+<script src="../../plugins/iCheck/icheck.min.js"></script>
 
 <!-- page script -->
 <script>
-      /*
-    $('#tbl-empRecord').DataTable({
-      "paging": true,
-      "lengthChange": false,
-      "searching": false,
-      "ordering": true,
-      "info": true,
-      "autoWidth": false
+    $("#fac_close_btn").on('click', function(){
+        window.location.reload();
     });
-    */
 </script>