+++ pages/src/link-file/head.php
@@ -5,16 +5,16 @@
 
 <meta content="width=device-width, initial-scale=1, maximum-scale=1, user-scalable=no" name="viewport">
 
-<link rel="stylesheet" href="../../../bootstrap/css/bootstrap.min.css">
+<link rel="stylesheet" href="../../bootstrap/css/bootstrap.min.css">
 
-<link rel="stylesheet" href="../../../dist/download/font-awesome-4.7.0/css/font-awesome.min.css">
+<link rel="stylesheet" href="../../dist/download/font-awesome-4.7.0/css/font-awesome.min.css">
 
-<link rel="stylesheet" href="../../../dist/download/ionicons-2.0.1/css/ionicons.min.css">
+<link rel="stylesheet" href="../../dist/download/ionicons-2.0.1/css/ionicons.min.css">
 
-<link rel="stylesheet" href="../../../dist/css/AdminLTE.min.css">
+<link rel="stylesheet" href="../../dist/css/AdminLTE.min.css">
 
-<link rel="stylesheet" href="../../../dist/css/skins/_all-skins.min.css">
+<link rel="stylesheet" href="../../dist/css/skins/_all-skins.min.css">
 
-<link rel="stylesheet" href="../../../dist/css/AdminLTE.min.css">
+<link rel="stylesheet" href="../../dist/css/AdminLTE.min.css">
 
-<link rel="stylesheet" href="../../../plugins/iCheck/square/blue.css">
+<link rel="stylesheet" href="../../plugins/iCheck/square/blue.css">