+++ pages/src/link-file/side-bar.php
@@ -70,6 +70,12 @@
                     <span>Year Level</span>
                 </a>
             </li>
+           <!--  <li class="treeview side1" id="backup">
+                <a href="backup">
+                    <i class="fa fa-users"></i> 
+                    <span>Backup Database</span>
+                </a>
+            </li> -->
         </ul>
     </section>
     <!-- /.sidebar -->