+++ pages/src/routes/classRoute.php
@@ -19,6 +19,28 @@ $route->url("class/add_class", function(){
 	$response = $class->add_class($data);
 	echo json_encode($response);
 });
+$route->url("class/update_class", function(){
+	define("RESTRICTED", true);
+	require_once "route_auth.php";
+	$class = new ClassController();
+	$data = array(
+		"yearlevel_id" => $_POST['yearlevel_id'],
+		"className" => $_POST['classname'],
+		"class_id" => (int)$_POST['class_id']
+	);
+	$response = $class->update_class($data);
+	echo json_encode($response);
+});
+$route->url("class/delete_class", function(){
+	define("RESTRICTED", true);
+	require_once "route_auth.php";
+	$data = array(
+		"class_id" => (int)$_GET['class_id']
+	);
+	$class = new ClassController();
+	$response = $class->delete_class($data);
+	echo json_encode($response);
+});
 $route->url("class/get_classes", function(){
 	define("RESTRICTED", true);
 	require_once "route_auth.php";